feat: document class function parameters inline
It is common to pass complex tables in as a function parameter. LuaLS
annotations don't provide nice-looking ways to type this, so often these
will get defined as classes elsewhere in the file (or in related files).
This change provides more utilities for parsing all lua types across a
project to get these class definitions and then insert them inline into
the generated function docs.